feat: add generic fetch params
