+++ src/components/SystemsView/index.ts
@@ -3,9 +3,8 @@ export type {
   SystemsViewProps,
   UseSystemsViewDataQuery,
   SystemsViewDataQueryResult,
-  SortDirection,
 } from './SystemsView';
-export type { SystemsViewFetchParams } from '../InventoryViews/hooks/useHostsQuery';
+export type { SystemsViewFetchParams, SortDirection } from './types';
 export type { OnInvalidate } from './SystemActionModalsContext';
 export type { SystemsViewActiveState } from './utils/deriveActiveState';
 export type { Column } from './columns/allColumnDefinitions';
