+++ src/components/SystemsView/types.ts
@@ -1,0 +1,13 @@
+import type { ISortBy } from '@patternfly/react-table';
+
+export type SortDirection = ISortBy['direction'];
+
+export type SystemsViewFetchParams<
+  TFilters extends Record<string, unknown> = Record<string, unknown>,
+> = {
+  page: number;
+  perPage: number;
+  sortBy: string | undefined;
+  direction: SortDirection | undefined;
+  filters: TFilters;
+};
